requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
html5lib==1.1
idna==3.4
importlib-metadata==7.0.1
lxml==4.9.3
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.3.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
thefuzz==0.20.0
tzdata==2023.3
urllib3==2.0.6
webencodings==0.5.1
zipp==3.17.0
